# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Real-time Chat API",
        "description": "Real-time chat application API",
        "contact": {
            "email": "contact@test.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor Local de Desenvolvimento"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/chats": {
            "get": {
                "tags": [
                    "Chats"
                ],
                "summary": "List all user chats",
                "operationId": "282fdd00e50391b0857b55cc3769325b",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page (default: 20)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of chats successfully retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ChatListItem"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/chats/direct": {
            "post": {
                "tags": [
                    "Chats"
                ],
                "summary": "Get or create a direct chat",
                "operationId": "6e9762fac3af8cc7056d2bd3cc2e7637",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "ID of the user to chat with",
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Chat retrieved or created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "chat": {
                                            "$ref": "#/components/schemas/ChatListItem"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error or cannot chat with self",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cannot create chat with yourself"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/chats/{id}": {
            "get": {
                "tags": [
                    "Chats"
                ],
                "summary": "Get a specific chat by ID",
                "operationId": "7fb2951546973e68c2290d294fbddb53",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the chat",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chat found successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "chat": {
                                            "$ref": "#/components/schemas/ChatListItem"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Chat not found or user not a member"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Chats"
                ],
                "summary": "Leave or delete a chat",
                "operationId": "99a2ee39c6b08e32fd1636a7ebf274eb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the chat",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Chat successfully left or deleted (No Content)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Chat deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Chat not found or user not a member"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/chats/{id}/read": {
            "put": {
                "tags": [
                    "Chats"
                ],
                "summary": "Mark all messages in a chat as read",
                "operationId": "cd4818bd8321e31ddbdb571840369a58",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the chat",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chat marked as read successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Chat marked as read"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Chat not found or user not a member"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ChatListUser": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 2
                    },
                    "name": {
                        "type": "string",
                        "example": "Alice Smith"
                    },
                    "email": {
                        "type": "string",
                        "example": "alice@example.com"
                    },
                    "avatar": {
                        "type": "string",
                        "example": "null",
                        "nullable": true
                    },
                    "is_online": {
                        "type": "boolean",
                        "example": true
                    },
                    "last_seen_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ChatListItem": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Chat"
                    },
                    {
                        "properties": {
                            "users": {
                                "description": "Outros usuários no chat.",
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/ChatListUser"
                                }
                            },
                            "last_message": {
                                "oneOf": [
                                    {
                                        "$ref": "#/components/schemas/Message"
                                    }
                                ],
                                "nullable": true,
                                "type": "object"
                            },
                            "unread_count": {
                                "description": "Número de mensagens não lidas.",
                                "type": "integer",
                                "example": 5
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "Chat": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "direct",
                            "group"
                        ],
                        "example": "direct"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Message": {
                "title": "Message",
                "required": [
                    "id",
                    "content",
                    "user_id",
                    "chat_id",
                    "created_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "content": {
                        "type": "string",
                        "example": "This is a message content."
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "chat_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-01T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-01T12:34:56Z"
                    },
                    "user": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "John Doe"
                            },
                            "avatar": {
                                "type": "string",
                                "format": "url",
                                "example": "http://example.com/avatar.jpg"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}